Refactor: image 리팩토링
image 패키지 리팩토링
errorcode 최상단으로 분리
domain -> application layer 네이밍 수정
dto 작성
