+++ src/main/java/com/comeon/backend/image/application/ImageDto.java
@@ -1,0 +1,23 @@
+package com.comeon.backend.image.application;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class ImageDto {
+
+    @Getter
+    @AllArgsConstructor
+    public static class UploadResponse {
+
+        private String imageUrl;
+    }
+
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class RemoveRequest {
+
+        private String imageUrl;
+    }
+}
