--- src/main/java/com/comeon/backend/image/domain/FileManager.java
@@ -1,10 +1,0 @@
-package com.comeon.backend.image.domain;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface FileManager {
-
-    String upload(MultipartFile multipartFile, Long userId);
-
-    void delete(String imageUrl);
-}
