+++ src/main/java/com/comeon/backend/image/presentation/ImageApiController.java
@@ -1,10 +1,9 @@
 package com.comeon.backend.image.presentation;
 
 import com.comeon.backend.common.security.JwtPrincipal;
-import com.comeon.backend.image.domain.FileManager;
-import com.comeon.backend.image.presentation.request.ImageRemoveRequest;
+import com.comeon.backend.image.application.FileManager;
+import com.comeon.backend.image.application.ImageDto;
 import com.comeon.backend.image.presentation.response.ImageRemoveResponse;
-import com.comeon.backend.image.presentation.response.ImageUploadResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -20,14 +19,14 @@ public class ImageApiController {
     private final FileManager fileManager;
 
     @PostMapping
-    public ImageUploadResponse uploadImage(@AuthenticationPrincipal JwtPrincipal jwtPrincipal,
+    public ImageDto.UploadResponse uploadImage(@AuthenticationPrincipal JwtPrincipal jwtPrincipal,
                                            @RequestParam("image") MultipartFile imageFile) {
-        return new ImageUploadResponse(fileManager.upload(imageFile, jwtPrincipal.getUserId()));
+        return fileManager.upload(imageFile, jwtPrincipal.getUserId());
     }
 
     @DeleteMapping
-    public ImageRemoveResponse removeImage(@RequestBody ImageRemoveRequest request) {
-        fileManager.delete(request.getImageUrl());
+    public ImageRemoveResponse removeImage(@RequestBody ImageDto.RemoveRequest request) {
+        fileManager.delete(request);
         return new ImageRemoveResponse();
     }
 }
