--- src/main/java/com/comeon/backend/image/presentation/request/ImageRemoveRequest.java
@@ -1,13 +1,0 @@
-package com.comeon.backend.image.presentation.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ImageRemoveRequest {
-
-    private String imageUrl;
-}
