--- src/main/java/com/comeon/backend/image/presentation/response/ImageUploadResponse.java
@@ -1,11 +1,0 @@
-package com.comeon.backend.image.presentation.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ImageUploadResponse {
-
-    private String imageUrl;
-}
